Remove unnecessary callback
The main reason for this is that spree_i18n adds the ability to add locale data to products through the admin. This is done on a separate page from the product edit page. This means that when we update the product's locale data, the store ids are not passed in, and the product is consequently removed from all its stores.